refactor: reorganize component structure and update import paths
Restructured the project by moving components into appropriate directories (e.g., `sections`, `common`, `modals`, `tables`) to improve maintainability. Updated import paths across the application to reflect the new structure. This refactor enhances code organization and makes it easier to locate and manage components.
