+++ src/components/tables/invoicesTable.jsx
@@ -1,41 +1,39 @@
 import React, { useState, useMemo } from "react";
 import DataTable from "react-data-table-component";
-import { Eye, Printer, Pencil } from "lucide-react";
-import { pdf } from "@react-pdf/renderer";
-import CotizacionPDF from "./cotizacionPDF";
-import ModalCreateQuote from "./modalSaveQuote";
+import { Download, Printer, Pencil } from "lucide-react";
+import ModalCreateInvoice from "../modals/modalSaveInvoice";
 
-const originalQuotes = [
+const invoicesData = [
   {
-    id: "COT-2025-001",
+    doc: "FAC-2025-01.pdf",
     cliente: "Gasera del Norte",
     fecha: "10/04/2023",
-    total: "$12,450.00",
-    estatus: "Aceptada",
+    pedido: "PED-2025-001",
+    estatus: "Completado",
     color: "green",
   },
   {
-    id: "COT-2025-002",
+    doc: "FAC-2025-02.pdf",
     cliente: "Gasolinera Sureste",
     fecha: "09/04/2023",
-    total: "$8,720.00",
+    pedido: "PED-2025-002",
     estatus: "En proceso",
     color: "yellow",
   },
   {
-    id: "COT-2025-003",
+    doc: "FAC-2025-03.pdf",
     cliente: "Distribuidora Central",
     fecha: "08/04/2023",
-    total: "$15,300.00",
-    estatus: "Rechazada",
+    pedido: "PED-2025-003",
+    estatus: "Atrasado",
     color: "red",
   },
   {
-    id: "COT-2025-004",
+    doc: "FAC-2025-04.pdf",
     cliente: "Gasera del Norte",
     fecha: "10/04/2023",
-    total: "$12,450.00",
-    estatus: "Aceptada",
+    pedido: "PED-2025-004",
+    estatus: "Completado",
     color: "green",
   },
 ];
@@ -46,32 +44,45 @@ const badgeColor = {
   red: "bg-red-100 text-red-700",
 };
 
-export default function QuotesTable() {
-  const [modalOpen, setModalOpen] = useState(false);
+export default function InvoicesTable() {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [invoices, setInvoices] = useState(invoicesData);
 
   const [searchText, setSearchText] = useState("");
   const [selectedStatus, setSelectedStatus] = useState("");
   const [selectedClient, setSelectedClient] = useState("");
 
-  const handleSaveQuote = (newQuote) => {
-    console.log("Nueva cotización:", newQuote);
+  const handleSave = (newInvoice) => {
+    setInvoices((prev) => [...prev, newInvoice]);
   };
 
-  const handleDownloadPDF = async (quote) => {
-    const blob = await pdf(<CotizacionPDF data={quote} />).toBlob();
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = `${quote.id}.pdf`;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
+  const filteredInvoices = useMemo(() => {
+    return invoices.filter((inv) => {
+      const matchesSearch =
+        inv.doc.toLowerCase().includes(searchText.toLowerCase()) ||
+        inv.pedido.toLowerCase().includes(searchText.toLowerCase());
+
+      const matchesStatus = selectedStatus
+        ? inv.estatus === selectedStatus
+        : true;
+
+      const matchesClient = selectedClient
+        ? inv.cliente === selectedClient
+        : true;
+
+      return matchesSearch && matchesStatus && matchesClient;
+    });
+  }, [searchText, selectedStatus, selectedClient, invoices]);
 
   const columns = [
     {
-      name: "ID Cotización",
-      selector: (row) => row.id,
+      name: "Documento",
+      selector: (row) => row.doc,
+      cell: (row) => (
+        <span className="text-blue-600 underline cursor-pointer">
+          {row.doc}
+        </span>
+      ),
       sortable: true,
     },
     {
@@ -84,8 +95,8 @@ export default function QuotesTable() {
       selector: (row) => row.fecha,
     },
     {
-      name: "Total",
-      selector: (row) => row.total,
+      name: "Pedido",
+      selector: (row) => row.pedido,
     },
     {
       name: "Estatus",
@@ -100,38 +111,25 @@ export default function QuotesTable() {
     {
       name: "Acciones",
       cell: (row) => (
-        <div className="flex items-center justify-start gap-3 h-full">
-          <Eye className="w-4 h-4 text-blue-500 cursor-pointer hover:scale-110 transition" />
-          <Printer
-            className="w-4 h-4 text-gray-600 cursor-pointer hover:scale-110 transition"
-            onClick={() => handleDownloadPDF(row)}
-          />
+        <div className="flex items-center gap-3">
+          <Download className="w-4 h-4 text-red-500 cursor-pointer hover:scale-110 transition" />
+          <Printer className="w-4 h-4 text-gray-600 cursor-pointer hover:scale-110 transition" />
           <Pencil className="w-4 h-4 text-yellow-500 cursor-pointer hover:scale-110 transition" />
         </div>
       ),
     },
   ];
 
-  // 🔍 Filtrado combinado con useMemo
-  const filteredQuotes = useMemo(() => {
-    return originalQuotes.filter((quote) => {
-      const matchesSearch = quote.id.toLowerCase().includes(searchText.toLowerCase());
-      const matchesStatus = selectedStatus ? quote.estatus === selectedStatus : true;
-      const matchesClient = selectedClient ? quote.cliente === selectedClient : true;
-      return matchesSearch && matchesStatus && matchesClient;
-    });
-  }, [searchText, selectedStatus, selectedClient]);
-
   return (
     <div className="bg-white p-6 rounded-xl shadow-md">
       {/* Header */}
       <div className="flex flex-col md:flex-row justify-between items-start md:items-center mb-6 gap-3">
-        <h2 className="text-xl font-bold">Gestión de Cotizaciones</h2>
+        <h2 className="text-xl font-bold">Gestión de Facturas y Remisiones</h2>
         <button
-          onClick={() => setModalOpen(true)}
+          onClick={() => setIsModalOpen(true)}
           className="bg-red-500 hover:bg-red-600 text-white px-4 py-2 rounded-md text-sm flex items-center gap-2"
         >
-          <span className="text-lg font-bold">+</span> Nueva Cotización
+          <span className="text-lg font-bold">+</span> Nueva Factura
         </button>
       </div>
 
@@ -139,7 +137,7 @@ export default function QuotesTable() {
       <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-4">
         <input
           type="text"
-          placeholder="Buscar cotización..."
+          placeholder="Buscar documento o pedido..."
           className="border border-gray-300 px-3 py-2 rounded-md text-sm w-full"
           value={searchText}
           onChange={(e) => setSearchText(e.target.value)}
@@ -150,9 +148,9 @@ export default function QuotesTable() {
           onChange={(e) => setSelectedStatus(e.target.value)}
         >
           <option value="">Filtrar por estatus</option>
-          <option value="Aceptada">Aceptada</option>
+          <option value="Completado">Completado</option>
           <option value="En proceso">En proceso</option>
-          <option value="Rechazada">Rechazada</option>
+          <option value="Atrasado">Atrasado</option>
         </select>
         <select
           className="border border-gray-300 px-3 py-2 rounded-md text-sm w-full"
@@ -166,10 +164,10 @@ export default function QuotesTable() {
         </select>
       </div>
 
-      {/* Tabla */}
+      {/* Tabla con filtros activos */}
       <DataTable
         columns={columns}
-        data={filteredQuotes}
+        data={filteredInvoices}
         pagination
         paginationComponentOptions={{
           rowsPerPageText: "Filas por página",
@@ -180,11 +178,11 @@ export default function QuotesTable() {
         striped
       />
 
-      {/* Modal de creación */}
-      <ModalCreateQuote
-        isOpen={modalOpen}
-        onClose={() => setModalOpen(false)}
-        onSave={handleSaveQuote}
+      {/* Modal */}
+      <ModalCreateInvoice
+        isOpen={isModalOpen}
+        onClose={() => setIsModalOpen(false)}
+        onSave={handleSave}
       />
     </div>
   );
